build.gradle.kts update. convert dawinMain to appleMain
