+++ src/appleMain/kotlin/com/configcat/NumberFormatter.kt
