+++ src/appleMain/kotlin/com/configcat/fetch/HttpRequestBuilder.kt
@@ -1,10 +1,10 @@
 package com.configcat.fetch
 
-import io.ktor.client.request.*
+import io.ktor.client.request.HttpRequestBuilder
 
 internal actual fun httpRequestBuilder(
     configCatUserAgent: String,
-    eTag: String
+    eTag: String,
 ): HttpRequestBuilder {
     return commonHttpRequestBuilder(configCatUserAgent, eTag)
 }
