+++ src/appleTest/kotlin/com/configcat/DarwinEvaluationTests.kt
@@ -4,12 +4,13 @@ import com.configcat.SingleValueCache
 import com.configcat.data.DarwinComparatorsTests
 import com.configcat.data.DarwinEpochDateValidationTests
 import com.configcat.evaluation.EvaluationTestLogger
-import com.configcat.evaluation.data.*
+import com.configcat.evaluation.data.TestSet
 import com.configcat.log.LogLevel
 import com.configcat.manualPoll
-import io.ktor.client.engine.mock.*
-import io.ktor.http.*
-import io.ktor.util.*
+import io.ktor.client.engine.mock.MockEngine
+import io.ktor.client.engine.mock.respond
+import io.ktor.http.HttpStatusCode
+import io.ktor.http.headersOf
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
 import kotlin.test.Test
@@ -17,18 +18,19 @@ import kotlin.test.fail
 
 @OptIn(ExperimentalCoroutinesApi::class)
 class DarwinEvaluationTests {
+    @Test
+    fun testComparators() =
+        runTest {
+            // The test contains formatted double value, which is different in case of JS module
+            testEvaluation(DarwinComparatorsTests)
+        }
 
     @Test
-    fun testComparators() = runTest {
-        // The test contains formatted double value, which is different in case of JS module
-        testEvaluation(DarwinComparatorsTests)
-    }
-
-    @Test
-    fun testEpochDateValidation() = runTest {
-        // The test contains formatted double value, which is different in case of JS module
-        testEvaluation(DarwinEpochDateValidationTests)
-    }
+    fun testEpochDateValidation() =
+        runTest {
+            // The test contains formatted double value, which is different in case of JS module
+            testEvaluation(DarwinEpochDateValidationTests)
+        }
 
     private suspend fun testEvaluation(testSet: TestSet) {
         var sdkKey = testSet.sdkKey
@@ -36,25 +38,27 @@ class DarwinEvaluationTests {
             sdkKey = TEST_SDK_KEY
         }
 
-        val mockEngine = MockEngine {
-            respond(
-                content = testSet.jsonOverride,
-                status = HttpStatusCode.OK,
-                headersOf(Pair("ETag", listOf("fakeETag")))
-            )
-        }
+        val mockEngine =
+            MockEngine {
+                respond(
+                    content = testSet.jsonOverride,
+                    status = HttpStatusCode.OK,
+                    headersOf(Pair("ETag", listOf("fakeETag"))),
+                )
+            }
 
         val evaluationTestLogger = EvaluationTestLogger()
-        val client = ConfigCatClient(sdkKey) {
-            pollingMode = manualPoll()
-            baseUrl = testSet.baseUrl
-            httpEngine = mockEngine
-            logger = evaluationTestLogger
-            logLevel = LogLevel.INFO
+        val client =
+            ConfigCatClient(sdkKey) {
+                pollingMode = manualPoll()
+                baseUrl = testSet.baseUrl
+                httpEngine = mockEngine
+                logger = evaluationTestLogger
+                logLevel = LogLevel.INFO
 
-            // add empty SingleValueCache to avoid JS extra cache logs
-            configCache = SingleValueCache("")
-        }
+                // add empty SingleValueCache to avoid JS extra cache logs
+                configCache = SingleValueCache("")
+            }
         client.forceRefresh()
 
         val tests = testSet.tests
