+++ src/appleTest/kotlin/com/configcat/data/DarwinComparatorsTests.kt
@@ -1,0 +1,1462 @@
+package com.configcat.data
+
+import com.configcat.ConfigCatUser
+import com.configcat.evaluation.data.TestCase
+import com.configcat.evaluation.data.TestSet
+
+object DarwinComparatorsTests : TestSet {
+    override val sdkKey = "configcat-sdk-1/JcPbCGl_1E-K9M-fJOyKyQ/OfQqcTjfFUGBwMKqtyEOrQ"
+    override val baseUrl = null
+    override val jsonOverride =
+        """
+                {
+           "p":{
+              "u":"https://cdn-global.configcat.com",
+              "r":0,
+              "s":"JEl\u002BhoGfr/01JCnpxr7kOCIoB2bYAM3uTMShm6HiAc4="
+           },
+           "f":{
+              "allinone":{
+                 "t":1,
+                 "r":[
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":20,
+                                "s":"0740e9103e26acb85b086dd9a15eaa84246902f096655371a5f001b9e13754e8"
+                             }
+                          },
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":21,
+                                "s":"0740e9103e26acb85b086dd9a15eaa84246902f096655371a5f001b9e13754e8"
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"1h"
+                          },
+                          "i":"e3a79156"
+                       }
+                    },
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":28,
+                                "s":"joe@example.com"
+                             }
+                          },
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":29,
+                                "s":"joe@example.com"
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"1c"
+                          },
+                          "i":"ed60451a"
+                       }
+                    },
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":16,
+                                "l":[
+                                   "0740e9103e26acb85b086dd9a15eaa84246902f096655371a5f001b9e13754e8"
+                                ]
+                             }
+                          },
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":17,
+                                "l":[
+                                   "0740e9103e26acb85b086dd9a15eaa84246902f096655371a5f001b9e13754e8"
+                                ]
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"2h"
+                          },
+                          "i":"aa24b7a3"
+                       }
+                    },
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":0,
+                                "l":[
+                                   "joe@example.com"
+                                ]
+                             }
+                          },
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":1,
+                                "l":[
+                                   "joe@example.com"
+                                ]
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"2c"
+                          },
+                          "i":"d37425a1"
+                       }
+                    },
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":22,
+                                "l":[
+                                   "4_e99c716658ca0b1035394161a3ca54f8dc688930ad90bed26aeff075cb947397"
+                                ]
+                             }
+                          },
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":23,
+                                "l":[
+                                   "4_e99c716658ca0b1035394161a3ca54f8dc688930ad90bed26aeff075cb947397"
+                                ]
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"3h"
+                          },
+                          "i":"5e6e0c6c"
+                       }
+                    },
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":30,
+                                "l":[
+                                   "joe@"
+                                ]
+                             }
+                          },
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":31,
+                                "l":[
+                                   "joe@"
+                                ]
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"3c"
+                          },
+                          "i":"5f562a70"
+                       }
+                    },
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":24,
+                                "l":[
+                                   "12_29030906a5c2729247ccad10154b56b84d61ee4d732361e0ba7c3817da4f91b3"
+                                ]
+                             }
+                          },
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":25,
+                                "l":[
+                                   "12_29030906a5c2729247ccad10154b56b84d61ee4d732361e0ba7c3817da4f91b3"
+                                ]
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"4h"
+                          },
+                          "i":"91b91d69"
+                       }
+                    },
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":32,
+                                "l":[
+                                   "@example.com"
+                                ]
+                             }
+                          },
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":33,
+                                "l":[
+                                   "@example.com"
+                                ]
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"4c"
+                          },
+                          "i":"4c80a977"
+                       }
+                    },
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":2,
+                                "l":[
+                                   "e@e"
+                                ]
+                             }
+                          },
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":3,
+                                "l":[
+                                   "e@e"
+                                ]
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"5"
+                          },
+                          "i":"dd12c429"
+                       }
+                    },
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Version",
+                                "c":4,
+                                "l":[
+                                   "1.0.0"
+                                ]
+                             }
+                          },
+                          {
+                             "u":{
+                                "a":"Version",
+                                "c":5,
+                                "l":[
+                                   "1.0.0"
+                                ]
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"6"
+                          },
+                          "i":"dba5d266"
+                       }
+                    },
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Version",
+                                "c":6,
+                                "s":"1.0.1"
+                             }
+                          },
+                          {
+                             "u":{
+                                "a":"Version",
+                                "c":9,
+                                "s":"1.0.1"
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"7"
+                          },
+                          "i":"1637ffc5"
+                       }
+                    },
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Version",
+                                "c":8,
+                                "s":"0.9.9"
+                             }
+                          },
+                          {
+                             "u":{
+                                "a":"Version",
+                                "c":7,
+                                "s":"0.9.9"
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"8"
+                          },
+                          "i":"b084ddd6"
+                       }
+                    },
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Number",
+                                "c":10,
+                                "d":1
+                             }
+                          },
+                          {
+                             "u":{
+                                "a":"Number",
+                                "c":11,
+                                "d":1
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"9"
+                          },
+                          "i":"d1d537a6"
+                       }
+                    },
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Number",
+                                "c":12,
+                                "d":1.1
+                             }
+                          },
+                          {
+                             "u":{
+                                "a":"Number",
+                                "c":15,
+                                "d":1.1
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"10"
+                          },
+                          "i":"52c846d0"
+                       }
+                    },
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Number",
+                                "c":14,
+                                "d":0.9
+                             }
+                          },
+                          {
+                             "u":{
+                                "a":"Number",
+                                "c":13,
+                                "d":0.9
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"11"
+                          },
+                          "i":"c91ffb7c"
+                       }
+                    },
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Date",
+                                "c":18,
+                                "d":1693497600
+                             }
+                          },
+                          {
+                             "u":{
+                                "a":"Date",
+                                "c":19,
+                                "d":1693497600
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"12"
+                          },
+                          "i":"c12182ef"
+                       }
+                    },
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Country",
+                                "c":26,
+                                "l":[
+                                   "5a85699e7343a36d89ee75dca859f7a73cb6be89182095bffb021d1d78de046c"
+                                ]
+                             }
+                          },
+                          {
+                             "u":{
+                                "a":"Country",
+                                "c":27,
+                                "l":[
+                                   "5a85699e7343a36d89ee75dca859f7a73cb6be89182095bffb021d1d78de046c"
+                                ]
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"13h"
+                          },
+                          "i":"a16b1a17"
+                       }
+                    },
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Country",
+                                "c":34,
+                                "l":[
+                                   "USA"
+                                ]
+                             }
+                          },
+                          {
+                             "u":{
+                                "a":"Country",
+                                "c":35,
+                                "l":[
+                                   "USA"
+                                ]
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"13c"
+                          },
+                          "i":"1a17d1b3"
+                       }
+                    }
+                 ],
+                 "v":{
+                    "s":"default"
+                 },
+                 "i":"9ff25f81"
+              },
+              "arrayContainsCaseCheckDogDefaultCat":{
+                 "t":1,
+                 "r":[
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Custom1",
+                                "c":26,
+                                "l":[
+                                   "00083f86e0f648b23f6721d43033bbef14378266fbf7de8a6760cc2ad237e9f3"
+                                ]
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"Dog"
+                          },
+                          "i":"5d80eff1"
+                       }
+                    }
+                 ],
+                 "v":{
+                    "s":"Cat"
+                 },
+                 "i":"ce055a38"
+              },
+              "arrayContainsDogDefaultCat":{
+                 "t":1,
+                 "r":[
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Custom1",
+                                "c":26,
+                                "l":[
+                                   "c2d5024661bc0e13f769cbb28bbfab7b78dac88b7876f007020f2e7cd47b1114"
+                                ]
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"Dog"
+                          },
+                          "i":"147fdd01"
+                       }
+                    }
+                 ],
+                 "v":{
+                    "s":"Cat"
+                 },
+                 "i":"5f573f9c"
+              },
+              "arrayDoesNotContainCaseCheckDogDefaultCat":{
+                 "t":1,
+                 "r":[
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Custom1",
+                                "c":27,
+                                "l":[
+                                   "14748140c64ecab48c7fd13f03811fe6390c8f578c99df96cf36fc2c6152f660"
+                                ]
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"Dog"
+                          },
+                          "i":"d4ad5730"
+                       }
+                    }
+                 ],
+                 "v":{
+                    "s":"Cat"
+                 },
+                 "i":"df4915fd"
+              },
+              "arrayDoesNotContainDogDefaultCat":{
+                 "t":1,
+                 "r":[
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Custom1",
+                                "c":27,
+                                "l":[
+                                   "bc0c24462c5098e434c63c7fcc6343af5000a4e7affd309f365edd4ccb7f428b"
+                                ]
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"Dog"
+                          },
+                          "i":"c2161ac9"
+                       }
+                    }
+                 ],
+                 "v":{
+                    "s":"Cat"
+                 },
+                 "i":"41910880"
+              },
+              "boolTrueIn202304":{
+                 "t":0,
+                 "r":[
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Custom1",
+                                "c":19,
+                                "d":1680307200
+                             }
+                          },
+                          {
+                             "u":{
+                                "a":"Custom1",
+                                "c":18,
+                                "d":1682899200
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "b":true
+                          },
+                          "i":"6948d7cd"
+                       }
+                    }
+                 ],
+                 "v":{
+                    "b":false
+                 },
+                 "i":"ae2a09bd"
+              },
+              "countryPercentageAttribute":{
+                 "t":1,
+                 "a":"Country",
+                 "p":[
+                    {
+                       "p":50,
+                       "v":{
+                          "s":"Falcon"
+                       },
+                       "i":"2b05fd81"
+                    },
+                    {
+                       "p":50,
+                       "v":{
+                          "s":"Horse"
+                       },
+                       "i":"e28b6a82"
+                    }
+                 ],
+                 "v":{
+                    "s":"Chicken"
+                 },
+                 "i":"29bb6bbb"
+              },
+              "customPercentageAttribute":{
+                 "t":1,
+                 "a":"Custom1",
+                 "p":[
+                    {
+                       "p":50,
+                       "v":{
+                          "s":"Falcon"
+                       },
+                       "i":"3715712d"
+                    },
+                    {
+                       "p":50,
+                       "v":{
+                          "s":"Horse"
+                       },
+                       "i":"7b3542d5"
+                    }
+                 ],
+                 "v":{
+                    "s":"Chicken"
+                 },
+                 "i":"50466fb6"
+              },
+              "missingPercentageAttribute":{
+                 "t":1,
+                 "a":"NotFound",
+                 "r":[
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":24,
+                                "l":[
+                                   "14_902a42101e8b77851c98456b383fd959ed0f5aed5b919b4a623c8c756cf0c3ab"
+                                ]
+                             }
+                          }
+                       ],
+                       "p":[
+                          {
+                             "p":50,
+                             "v":{
+                                "s":"Falcon"
+                             },
+                             "i":"4b7d88ba"
+                          },
+                          {
+                             "p":50,
+                             "v":{
+                                "s":"Horse"
+                             },
+                             "i":"a1c2c9a9"
+                          }
+                       ]
+                    },
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":24,
+                                "l":[
+                                   "14_902a42101e8b77851c98456b383fd959ed0f5aed5b919b4a623c8c756cf0c3ab"
+                                ]
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"NotFound"
+                          },
+                          "i":"8aa042fe"
+                       }
+                    }
+                 ],
+                 "v":{
+                    "s":"Chicken"
+                 },
+                 "i":"e5107172"
+              },
+              "stringArrayContainsAnyOfCleartextDogDefaultCat":{
+                 "t":1,
+                 "r":[
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Custom1",
+                                "c":34,
+                                "l":[
+                                   "read",
+                                   "write",
+                                   "execute"
+                                ]
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"Dog"
+                          },
+                          "i":"9ddb8a37"
+                       }
+                    }
+                 ],
+                 "v":{
+                    "s":"Cat"
+                 },
+                 "i":"0d45ab4b"
+              },
+              "stringArrayContainsAnyOfDogDefaultCat":{
+                 "t":1,
+                 "r":[
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Custom1",
+                                "c":26,
+                                "l":[
+                                   "090415ce6b462a2152e06d68aeb7c452a564d19a262eb959c510636a189e105d",
+                                   "0aa50b49ca02a59cf507856d88ab25b76a7e69e553d25e67254359d8bbb8b1d6",
+                                   "1aa91982f7ccae1943f05ac437d635b3261e3ce06aba846dd2ecc6332e4675c2"
+                                ]
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"Dog"
+                          },
+                          "i":"aa03b1ff"
+                       }
+                    }
+                 ],
+                 "v":{
+                    "s":"Cat"
+                 },
+                 "i":"203317f5"
+              },
+              "stringArrayNotContainsAnyOfCleartextDogDefaultCat":{
+                 "t":1,
+                 "r":[
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Custom1",
+                                "c":35,
+                                "l":[
+                                   "read",
+                                   "write",
+                                   "execute"
+                                ]
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"Dog"
+                          },
+                          "i":"15c865df"
+                       }
+                    }
+                 ],
+                 "v":{
+                    "s":"Cat"
+                 },
+                 "i":"6df210da"
+              },
+              "stringArrayNotContainsAnyOfDogDefaultCat":{
+                 "t":1,
+                 "r":[
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Custom1",
+                                "c":27,
+                                "l":[
+                                   "b645a36d4d2e24a0612296ded074eace87ce10a0da21c5cb2ac9a6dccbe79cc5",
+                                   "552a92975423ea255384f48a6387be172c05ac72238f90050cdfe27cc4659cfd",
+                                   "4c00d4171202dbeb35830b1df8e47185968351c771bb2f633ea50ac1c049e016"
+                                ]
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"Dog"
+                          },
+                          "i":"259816ba"
+                       }
+                    }
+                 ],
+                 "v":{
+                    "s":"Cat"
+                 },
+                 "i":"60b961b0"
+              },
+              "stringContainsAnyOfDogDefaultCat":{
+                 "t":1,
+                 "r":[
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":2,
+                                "l":[
+                                   "@verizon.com",
+                                   "@verizon.net",
+                                   "@aol.com"
+                                ]
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"Dog"
+                          },
+                          "i":"09af657f"
+                       }
+                    }
+                 ],
+                 "v":{
+                    "s":"Cat"
+                 },
+                 "i":"063bcf39"
+              },
+              "stringDoseNotEqualDogDefaultCat":{
+                 "t":1,
+                 "r":[
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":21,
+                                "s":"3cb496a1c3844215c784116ce9e91c143860d7ef18f7fadadcc901b8df5d235c"
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"Dog"
+                          },
+                          "i":"8e423808"
+                       }
+                    }
+                 ],
+                 "v":{
+                    "s":"Cat"
+                 },
+                 "i":"1835a09a"
+              },
+              "stringEndsWithAnyOfCleartextDogDefaultCat":{
+                 "t":1,
+                 "r":[
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":32,
+                                "l":[
+                                   "@verizon.com",
+                                   "@verizon.net",
+                                   "@aol.com"
+                                ]
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"Dog"
+                          },
+                          "i":"33d35402"
+                       }
+                    }
+                 ],
+                 "v":{
+                    "s":"Cat"
+                 },
+                 "i":"31976ec3"
+              },
+              "stringEndsWithAnyOfDogDefaultCat":{
+                 "t":1,
+                 "r":[
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":24,
+                                "l":[
+                                   "12_781f5f9b054a14721a835fcdd2d03ac6d45d99eb55b387bf5938904c8f65aa35",
+                                   "12_d15f199cfbd96ef1c6f7683e66d5e0f85c9c591ce377289abb2e785fc71bbdf1",
+                                   "8_6ee234c513b7518bd705cc1049576c1e757a201200ff26a6a3a91821f954c6cb"
+                                ]
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"Dog"
+                          },
+                          "i":"7231ddf8"
+                       }
+                    }
+                 ],
+                 "v":{
+                    "s":"Cat"
+                 },
+                 "i":"de17fd2a"
+              },
+              "stringEndsWithDogDefaultCat":{
+                 "t":1,
+                 "r":[
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":24,
+                                "l":[
+                                   "14_236ec291c9b54fad3373a0b7e0e465b33459198fd1027838c101516ad8ae1b39"
+                                ]
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"Dog"
+                          },
+                          "i":"d7a00741"
+                       }
+                    }
+                 ],
+                 "v":{
+                    "s":"Cat"
+                 },
+                 "i":"45b7d922"
+              },
+              "stringEqualsCleartextDogDefaultCat":{
+                 "t":1,
+                 "r":[
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":28,
+                                "s":"a@configcat.com"
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"Dog"
+                          },
+                          "i":"087e01dd"
+                       }
+                    }
+                 ],
+                 "v":{
+                    "s":"Cat"
+                 },
+                 "i":"89785ab3"
+              },
+              "stringEqualsDogDefaultCat":{
+                 "t":1,
+                 "r":[
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":20,
+                                "s":"86e6b430939acac093f3d8c48be10798896f0abf3b96e2e39080acedd925d887"
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"Dog"
+                          },
+                          "i":"703c31ed"
+                       }
+                    }
+                 ],
+                 "v":{
+                    "s":"Cat"
+                 },
+                 "i":"adc0b01c"
+              },
+              "stringNotContainsAnyOfDogDefaultCat":{
+                 "t":1,
+                 "r":[
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":3,
+                                "l":[
+                                   "@verizon.com",
+                                   "@verizon.net",
+                                   "@aol.com"
+                                ]
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"Dog"
+                          },
+                          "i":"49627b36"
+                       }
+                    }
+                 ],
+                 "v":{
+                    "s":"Cat"
+                 },
+                 "i":"36848b03"
+              },
+              "stringNotEndsWithAnyOfCleartextDogDefaultCat":{
+                 "t":1,
+                 "r":[
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":33,
+                                "l":[
+                                   "@verizon.com",
+                                   "@verizon.net",
+                                   "@aol.com"
+                                ]
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"Dog"
+                          },
+                          "i":"886ced9d"
+                       }
+                    }
+                 ],
+                 "v":{
+                    "s":"Cat"
+                 },
+                 "i":"864b6202"
+              },
+              "stringNotEndsWithAnyOfDogDefaultCat":{
+                 "t":1,
+                 "r":[
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":25,
+                                "l":[
+                                   "12_295e66dc483034349bc6cffd14190ccff949ac1f34df5fbf01437b8162719b98",
+                                   "12_11933417446186b92f63db5931a275f156485d0aef5ee8e265afb350921bd0b1",
+                                   "8_aa181c1d9462e3916cb05669e2c408541edf03dad19a9655b340eb32ddd4d060"
+                                ]
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"Dog"
+                          },
+                          "i":"6eb0ec3a"
+                       }
+                    }
+                 ],
+                 "v":{
+                    "s":"Cat"
+                 },
+                 "i":"7020bcd6"
+              },
+              "stringNotEndsWithDogDefaultCat":{
+                 "t":1,
+                 "r":[
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":25,
+                                "l":[
+                                   "14_141d3f398885dd06d6c14e6629602125d8cb0dddf2ef85896f682c74abb4bc28"
+                                ]
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"Dog"
+                          },
+                          "i":"d37b6f18"
+                       }
+                    }
+                 ],
+                 "v":{
+                    "s":"Cat"
+                 },
+                 "i":"91ba1bcb"
+              },
+              "stringNotEqualsCleartextDogDefaultCat":{
+                 "t":1,
+                 "r":[
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":29,
+                                "s":"b@configcat.com"
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"Dog"
+                          },
+                          "i":"3ace20fb"
+                       }
+                    }
+                 ],
+                 "v":{
+                    "s":"Cat"
+                 },
+                 "i":"09c9725f"
+              },
+              "stringNotStartsWithAnyOfCleartextDogDefaultCat":{
+                 "t":1,
+                 "r":[
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":31,
+                                "l":[
+                                   "u@",
+                                   "a@",
+                                   "b@"
+                                ]
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"Dog"
+                          },
+                          "i":"3717f2ca"
+                       }
+                    }
+                 ],
+                 "v":{
+                    "s":"Cat"
+                 },
+                 "i":"1d661433"
+              },
+              "stringNotStartsWithAnyOfDogDefaultCat":{
+                 "t":1,
+                 "r":[
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":23,
+                                "l":[
+                                   "2_b0ab1063bba4431f95710a8bd9c9e2bcd1cebc09fe6803cf87ee501d045e8ee0",
+                                   "2_9206a0a6b987410488c8020650788c597509af05aa5327f92cae1c999c401c34",
+                                   "2_d8f5c200e2c54b0f84f7df024576f799f484d0e258c6058142a4993daa5e2998"
+                                ]
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"Dog"
+                          },
+                          "i":"b5ba025e"
+                       }
+                    }
+                 ],
+                 "v":{
+                    "s":"Cat"
+                 },
+                 "i":"c35929e3"
+              },
+              "stringNotStartsWithDogDefaultCat":{
+                 "t":1,
+                 "r":[
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":23,
+                                "l":[
+                                   "1_061a38ed8d5955d90b88d1b1949512a45032390a2581f3c7e36597f7459a48c7"
+                                ]
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"Dog"
+                          },
+                          "i":"72c4e1ac"
+                       }
+                    }
+                 ],
+                 "v":{
+                    "s":"Cat"
+                 },
+                 "i":"2b16da78"
+              },
+              "stringStartsWithAnyOfCleartextDogDefaultCat":{
+                 "t":1,
+                 "r":[
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":30,
+                                "l":[
+                                   "u@",
+                                   "a@",
+                                   "b@"
+                                ]
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"Dog"
+                          },
+                          "i":"9e55f5cf"
+                       }
+                    }
+                 ],
+                 "v":{
+                    "s":"Cat"
+                 },
+                 "i":"e170a185"
+              },
+              "stringStartsWithAnyOfDogDefaultCat":{
+                 "t":1,
+                 "r":[
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":22,
+                                "l":[
+                                   "2_52f47d1a193071a304e41812055cc1282d90287e5c993ae159fa08fb1ccd3656",
+                                   "2_6e5ae1bb586660088d330ac106bbf6b7f2b43be1ca70dbb766f203b76bc84049",
+                                   "2_3edba8b738135ad32e85d8695acb8d8511ac67d83f50da55abd9ba469da88efe"
+                                ]
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"Dog"
+                          },
+                          "i":"1d9b7603"
+                       }
+                    }
+                 ],
+                 "v":{
+                    "s":"Cat"
+                 },
+                 "i":"dd5b3211"
+              },
+              "stringStartsWithDogDefaultCat":{
+                 "t":1,
+                 "r":[
+                    {
+                       "c":[
+                          {
+                             "u":{
+                                "a":"Email",
+                                "c":22,
+                                "l":[
+                                   "1_55e0b0566d89dc0fda2323efcfb958c782a4648513bdcc4dc84b044fd34230dd"
+                                ]
+                             }
+                          }
+                       ],
+                       "s":{
+                          "v":{
+                             "s":"Dog"
+                          },
+                          "i":"3b409872"
+                       }
+                    }
+                 ],
+                 "v":{
+                    "s":"Cat"
+                 },
+                 "i":"3659b0fe"
+              }
+           }
+        }
+        """.trimIndent()
+    override val tests: Array<TestCase> =
+        arrayOf(
+            TestCase(
+                key = "allinone",
+                defaultValue = "",
+                returnValue = "default",
+                user =
+                    ConfigCatUser(
+                        "12345",
+                        "joe@example.com",
+                        "[\"USA\"]",
+                        custom = mapOf("Version" to "1.0.0", "Number" to "1.0", "Date" to "1693497500"),
+                    ),
+                expectedLog = """INFO [5000] Evaluating 'allinone' for User '{"Identifier":"12345","Email":"joe@example.com","Country":"[\"USA\"]","Version":"1.0.0","Number":"1.0","Date":"1693497500"}'
+  Evaluating targeting rules and applying the first match if any:
+  - IF User.Email EQUALS '<hashed value>' => true
+    AND User.Email NOT EQUALS '<hashed value>' => false, skipping the remaining AND conditions
+    THEN '1h' => no match
+  - IF User.Email EQUALS 'joe@example.com' => true
+    AND User.Email NOT EQUALS 'joe@example.com' => false, skipping the remaining AND conditions
+    THEN '1c' => no match
+  - IF User.Email IS ONE OF [<1 hashed value>] => true
+    AND User.Email IS NOT ONE OF [<1 hashed value>] => false, skipping the remaining AND conditions
+    THEN '2h' => no match
+  - IF User.Email IS ONE OF ['joe@example.com'] => true
+    AND User.Email IS NOT ONE OF ['joe@example.com'] => false, skipping the remaining AND conditions
+    THEN '2c' => no match
+  - IF User.Email STARTS WITH ANY OF [<1 hashed value>] => true
+    AND User.Email NOT STARTS WITH ANY OF [<1 hashed value>] => false, skipping the remaining AND conditions
+    THEN '3h' => no match
+  - IF User.Email STARTS WITH ANY OF ['joe@'] => true
+    AND User.Email NOT STARTS WITH ANY OF ['joe@'] => false, skipping the remaining AND conditions
+    THEN '3c' => no match
+  - IF User.Email ENDS WITH ANY OF [<1 hashed value>] => true
+    AND User.Email NOT ENDS WITH ANY OF [<1 hashed value>] => false, skipping the remaining AND conditions
+    THEN '4h' => no match
+  - IF User.Email ENDS WITH ANY OF ['@example.com'] => true
+    AND User.Email NOT ENDS WITH ANY OF ['@example.com'] => false, skipping the remaining AND conditions
+    THEN '4c' => no match
+  - IF User.Email CONTAINS ANY OF ['e@e'] => true
+    AND User.Email NOT CONTAINS ANY OF ['e@e'] => false, skipping the remaining AND conditions
+    THEN '5' => no match
+  - IF User.Version IS ONE OF ['1.0.0'] => true
+    AND User.Version IS NOT ONE OF ['1.0.0'] => false, skipping the remaining AND conditions
+    THEN '6' => no match
+  - IF User.Version < '1.0.1' => true
+    AND User.Version >= '1.0.1' => false, skipping the remaining AND conditions
+    THEN '7' => no match
+  - IF User.Version > '0.9.9' => true
+    AND User.Version <= '0.9.9' => false, skipping the remaining AND conditions
+    THEN '8' => no match
+  - IF User.Number = '1' => true
+    AND User.Number != '1' => false, skipping the remaining AND conditions
+    THEN '9' => no match
+  - IF User.Number < '1.1' => true
+    AND User.Number >= '1.1' => false, skipping the remaining AND conditions
+    THEN '10' => no match
+  - IF User.Number > '0.9' => true
+    AND User.Number <= '0.9' => false, skipping the remaining AND conditions
+    THEN '11' => no match
+  - IF User.Date BEFORE '1693497600' (2023-08-31T16:00:00.000Z UTC) => true
+    AND User.Date AFTER '1693497600' (2023-08-31T16:00:00.000Z UTC) => false, skipping the remaining AND conditions
+    THEN '12' => no match
+  - IF User.Country ARRAY CONTAINS ANY OF [<1 hashed value>] => true
+    AND User.Country ARRAY NOT CONTAINS ANY OF [<1 hashed value>] => false, skipping the remaining AND conditions
+    THEN '13h' => no match
+  - IF User.Country ARRAY CONTAINS ANY OF ['USA'] => true
+    AND User.Country ARRAY NOT CONTAINS ANY OF ['USA'] => false, skipping the remaining AND conditions
+    THEN '13c' => no match
+  Returning 'default'.""",
+            ),
+        )
+}
